requirements.txt
mysql-connector-python==8.0.33
flask-mail==0.9.1
python-dotenv==0.21.0
Flask==3.0.1
Flask-MySQLdb==0.2.0
WTForms==3.0.1
Flask-WTF==1.0.1
razorpay==1.3.0
pyotp==2.6.0
reportlab==3.6.12
blinker==1.7.0
click==8.1.7
Flask-Scss==0.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pyScss==1.4.0
Werkzeug==3.0.1
flask-wtf==1.1.1
Flask-Limiter==2.9.1
